+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.0
+ * Updated to `@animoca/ethereum-contracts-nft_staking:4.0.0`.
+ * Staking is separated in phase 1 and phase 2.
+ * Phase 2 requires escrowing of REVV for staking.
+
 ## 0.4.0
  * Added `NFTRepairCentre`, a contract to manage defunct NFTs.
  * Updated to `@animoca/ethereum-contracts-core_library:3.1.1`.